Publish UTM data as a infuse_msgs/asn1_bitstream instead of a infuse_novatel_gps_msgs/NovatelUtmPosition
This commit comments out all code used to fill and publish a
infuse_novatel_gps_msgs/NovatelUtmPosition and replaces it by
equivalent code to fill and publish a TransformWithCovariance encoded
in a infuse_msgs/asn1_bitstream.

The position is encoded as ENU and the orientation is kept fixed. The
covariances are filled from the ENU sigma values received from the
sensor.

parentFrameId is set to "GlobalTerrainFrame", childFrameId to
"RoverBodyFrame", and producerId to "novatel_gps_driver"

All the heavy magic is done on
novatel_gps_driver/src/parsers/bestutm.cpp